### 2021-08-24 ae02b0e
added pooch==1.4.0 to README env config
