### 2021-09-02 #3666 @crvernon
:mega:  Mid-week rally!  Just checking in to see how things are going @dbuscombe-usgs and @chrisleaman ?  Don't hesitate to reach out if you have any questions.

👏 Keep up the good work!
