### 2021-09-03 #3666 @chrisleaman
Thanks @crvernon, I've had a read of the paper and just summarizing my thoughts. Also planning to run the package on some of my own data to test it out 😄
