### 2021-09-03 npucino/sandpyper#6 @dbuscombe-usgs
[JOSS REVIEW] Install instructions
Installation instructions should be improved  - as currently written, the instructions are not an optimal way to ensure success across different platforms and for relatively novice users who don't necessarily know the pitfalls with using pip and conda together in this way ...

and, I cannot get them to work, because the pip repository (why?) is packaged with dependencies that break the conda environment. In fact, the installation is a mess for a number of reasons:

1. the instructions don't mention what version of python to use, which is crucial. You should recommend a version that you know to work across multiple platforms. I am initially testing on Windows, so my python version will naturally be ahead of my Linux computer

2. installing some dependencies prior to installing the conda environment is a VERY bad idea because they'll install into the base environment, so won't be accessible to the sandpyper environment, and also would very likely break the base environment

3. You should consider providing a conda yml file

4. Further, you are being very restrictive with these versions (that are not compatible with my conda install and OS, for example)

`conda install  geopandas=0.8.2 matplotlib=3.3.4 numpy=1.20.1 pandas=1.2.2 tqdm=4.56.2 pysal=2.1 rasterio=1.2.0 richdem=0.3.4 scikit-image=0.18.1 scikit-learn=0.24.1 scipy=1.6.0 seaborn=0.11.1 tqdm=4.56.2 pooch=1.4.0 fuzzywuzzy`

and that is bad way to 'future-proof' your installation (sorry, but you WILL be getting issues about this). Further, that's not a good way to use conda - conda is designed to tell YOU what versions YOU need, and the requirements.txt file would be the usual place to put really specific versions (for a pip install)

5. There is no need to install `richdem` and visual studio tools (thatmight require admin privileges). Conda will install vs build tools for you that for you if you DONT provide the specific version numbers. And `richdem` is installed by conda, so that's a confusing instruction to install that separately. You can therefore remove that confusing step.

6. the `jupyter` package should be installed with the rest, because of the very complex dependencies that might break conda with a posthoc installation. Also the package is called `jupyter`, not `jupyter notebook` (that is the command you run afterwards)

7. the pip packages `sandpyper` ships with versions of numpy, pandas, rasterio and matplotlib that either conflict with the conda environment, or should simply be installed with the conda environment. The  pip installation also raises GDAL dependency errors. That should be in the conda environment!

Conda and pip have very different purposes. I STRONGLY recommend you make a conda environment with all of the dependencies contained within. Its not at all clear to me why the package itself has to be a pip installation, but if you insist of that, it should ship with no additionally dependencies. At which point, there's no point making it a pip installation - you see my point? Its very confused.

I am therefore recommending the following approach, creating an environment with a specific python version

`conda create --name sandpyper_env python=3.7`

then installing ALL the dependencies inside it.

`conda install geopandas matplotlib numpy pandas tqdm pysal rasterio richdem scikit-image scikit-learn scipy seaborn tqdm pooch fuzzywuzzy jupyter gdal`

That approach a) was the only  thing that worked for me,  b) is superior for the reasons I describe above, and c) circumvents the need to install `richdem` and visual studio tools

Therefore the entire installation could be simply for following

`conda config --add channels conda-forge`

then

`conda create --name sandpyper_env python=3.7`

then

`conda activate sandpyper`

`conda install geopandas matplotlib numpy pandas tqdm pysal rasterio richdem scikit-image scikit-learn scipy seaborn tqdm pooch fuzzywuzzy gdal`

and finally

`pip install sandpyper`

(however, that results in GDAL dependency errors because pip/pypi is not appropriate here)

so, use pip/git

`pip install git+https://github.com/npucino/sandpyper.git`

that also gives GDAL errors

So, I cant install on windows. I cant review the code until I can install

I strongly recommend testing installation instructions on multiple machines before posting them
