### 2021-09-03 #3666 @dbuscombe-usgs
@crvernon @npucino I am having a lot of difficulty installing this package using the provided instructions. I have a detailed issue [here](https://github.com/npucino/sandpyper/issues/6)

I have a lot of experince with conda and this is just a bad way to install a package with all of these very complicated dependencies. I'm stalled out until I can get the program installed
