### 2021-09-03 npucino/sandpyper#6 @dbuscombe-usgs
I didnt install GDAL from a wheel like instructed because I happen to know that rasterio/gdal will install on my machine through conda. It is the pip package that is breaking GDAL
