### 2021-09-03 npucino/sandpyper#6 @dbuscombe-usgs
If a conda environment is created this way (with the package that has least chance of successful integration with the others, which is probably gdal)

`conda create --name sandpyper_env gdal`

`conda activate sandpyper_env`
`conda install geopandas matplotlib numpy pandas tqdm pysal rasterio richdem scikit-image scikit-learn scipy seaborn tqdm pooch fuzzywuzzy`

At this point. all the dependencies are installed correctly. It is then the pip package `pip install sandpyper` (with its really specific and conflicting dependencies) that break GDAL

I'll keep trying to make this work ...
