### 2021-09-03 npucino/sandpyper#6 @dbuscombe-usgs
Progress! I have a working solution for my windows machine. Conda works best when you tell it upfront everything you're asking it to install, including all dependencies and pip dependencies. Therefore I recommend going the route of installing from a yml file.

This yml file successfully installs the package and its dependencies . Paste the following into a file called `sandpyper.yml`
```
name: sandpyper_env
channels:
  - conda-forge
  - defaults
dependencies:
 - python
 - scipy
 - numpy
 - scikit-image
 - scikit-learn
 - seaborn
 - pooch
 - fuzzywuzzy
 - ipython
 - jupyter
 - rasterio
 - tqdm
 - geopandas
 - matplotlib
 - richdem
 - pip
 - pip:
   - sandpyper
```

then, install using

`conda env create --file sandpyper.yml`

That is successful, and also much easier to communicate

however, when I `import sandpyper` there is a syntax error in the init file of the package

```
In [2]: import sandpyper
Traceback (most recent call last):

  File "C:\Users\dbuscombe\Anaconda3\envs\sandpyper_env\lib\site-packages\IPython\core\interactiveshell.py", line 3441, in run_code
    exec(code_obj, self.user_global_ns, self.user_ns)

  File "<ipython-input-2-bf14279135af>", line 1, in <module>
    import sandpyper

  File "C:\Users\dbuscombe\Anaconda3\envs\sandpyper_env\lib\site-packages\sandpyper\__init__.py", line 1
    __version__='__version__ = '0.0.2''
                                ^
SyntaxError: invalid syntax
```
