### 2021-09-03 #3666 @dbuscombe-usgs
I made my own yml file installer to deal with the issues of the mixed conda/pip installation, that works a lot better and in fewer lines of code. see [here](https://github.com/npucino/sandpyper/issues/6#issuecomment-912850421) However, the `__init__` script for the sandpyper pip package has broken syntax that means I get errors . So, I should now be able to run this software when @npucino gets a chance to revisit that.

Installation is typically the most difficult thing to get right with these types of python packages. I expect it will be plain sailing after that!
