### 2021-09-03 npucino/sandpyper#6 @dbuscombe-usgs
The fix is to change the first line of __init__.py in the installed package to

`__version__="__version__ = '0.0.2'"`

then `import sandpyper` gives another dependency error:

ModuleNotFoundError: No module named 'pysal'

luckily, conda has that:
`conda install pysal`

that install successfully. However, now I run into an issue with `pysal`, so I'm guessing the pysal version that conda installs is older?

I'm going to give up troubleshooting for now until @npucino has a chance to take a look at all of this

```
In [1]: import sandpyper
C:\Users\dbuscombe\Anaconda3\envs\sandpyper_env\lib\site-packages\numba\np\ufunc\parallel.py:366: NumbaWarning: The TBB threading layer requires TBB version 2021 update 1 or later i.e., TBB_INTERFACE_VERSION >= 12010. Found TBB_INTERFACE_VERSION = 11100. The TBB threading layer is disabled.
  warnings.warn(problem)
ImportError                               Traceback (most recent call last)
<ipython-input-1-bf14279135af> in <module>
----> 1 import sandpyper

~\Anaconda3\envs\sandpyper_env\lib\site-packages\sandpyper\__init__.py in <module>
      3 __email__ = 'npucino@deakin.edu.au'
      4
----> 5 from . import dynamics
      6 from . import hotspot
      7 from . import labels

~\Anaconda3\envs\sandpyper_env\lib\site-packages\sandpyper\dynamics\__init__.py in <module>
      1 __version__='0.0.1'
----> 2 from .dynamics import (
      3 attach_trs_geometry,
      4 infer_weights,
      5 get_coastal_Markov,

~\Anaconda3\envs\sandpyper_env\lib\site-packages\sandpyper\dynamics\dynamics.py in <module>
      4 import geopandas as gpd
      5
----> 6 from pysal.explore.giddy.markov import Markov
      7 import matplotlib.pyplot as plt
      8 import seaborn as sb

ImportError: cannot import name 'Markov' from 'pysal.explore.giddy.markov' (C:\Users\dbuscombe\Anaconda3\envs\sandpyper_env\lib\site-packages\pysal\explore\giddy\markov\__init__.py)
```
