### 2021-09-03 npucino/sandpyper#6 @dbuscombe-usgs
Updated yml file to deal with the pysal dependency:

```
name: sandpyper_env
channels:
  - conda-forge
  - defaults
dependencies:
 - python
 - scipy
 - numpy
 - scikit-image
 - scikit-learn
 - seaborn
 - pooch
 - fuzzywuzzy
 - ipython
 - jupyter
 - rasterio
 - tqdm
 - geopandas
 - matplotlib
 - richdem
 - pysal
 - pip
 - pip:
   - sandpyper
```
