### 2021-09-03 npucino/sandpyper#6 @dbuscombe-usgs
The pysal version is 2.5.0
