### 2021-09-03 #3666 @npucino
@dbuscombe-usgs thanks for the detailed issue report, I will fix this Monday. I tested the package installation and unit tested the code in github action and assumed it worked out in the real world.
Also I noticed that for some reason your pip install sandpyper downloaded a very old version 0.0.2 (with that typo in the init file), while the latest one in pip is 1.0.0 which is drastically different. I think it is becasue the pip install GDAL confilcts made pip to roll back to a previous version where the GDAL doesn't break the imports. I ll look into that!
