### 2021-09-03 #3666 @dbuscombe-usgs
Great, thanks for the quick response @npucino - that sounds like a good plan. I'll check back in Monday and try to help troubleshoot. I'm pretty good at finding conda and pip issues in other people's code, so don't take it personally (I also really hate the python ecosystem for its crummy package managers!)

in the yml file I made, it does successfully create all the dependencies in conda using conda-forge. then it is looking for a version of sandpyper on pip that is compatible, That is version 0.0.2. That would explain why, I think

but it does prove that a) you can use conda entirely to set up the dependencies in a conda environment without being overly prescriptive about specific versions of packages, and that b) its all the pip-only dependencies that are breaking the install. Its not clear why, for example, why the requirements.txt in the pypi package lists the same libraries as you suggest installing into the conda environment ....

my general recommendation here is that conda environments as a rule want to install conda packages, not pip ones. It works, but only if you allow conda to take precedence for as many of the dependencies as you can, and dont cross-post dependencies between conda and pip, you'll likely have fewer install dependencies

I can also try installing on linux and report back what I find on the issues tab on the repo page
