### 2021-09-04 #3666 @npucino
No no I do not take it personally @dbuscombe-usgs, please be as critic as you can I need it to grow and get better at this and improve that package! No mercy is good science. I think I will have to change the installation procedure and move it all to conda rather than pip. CHeers
