### 2021-09-06 npucino/sandpyper#6 @npucino
Thanks @dbuscombe-usgs and sorry for the numerus issues you encountered during installation.

I just tried the installation on a new Windows machine by following the guidelines I wrote in the README  and everything worked out smoothly with not a single errors. Which is not necessarily a good thing cause I have to understand what went wrong with yours.
Let's review your issues:

1.  I must write the supported Python versions and OS you are right, thanks for noting that.
2.  The only dependency I thought people will need to have prior the snadpyper environment is VS build tools, not Richdem. I see the confusion here. What I meant in the README is simply that you need to have VS build tools installed in your machine because Richdem will need it (during the conda installation). The fact that I put an hyperlink to Richdem must have misled you into thinking that Richdem must be installed prior the sandpyper_env in your base environment, which is definitely not what I meant and not what is intended. I change the sentence to make it more clear, thanks!
3. I agree that PIP + CONDA = MESS,  so one unique yml file would be ideal and also to package everything into conda would be much more efficient. So I do that asap. Hopefully there will not be the need to install  VS build tools separately.
4. Yes. I am being restrictive becasue those are the versions of the packages I am sure will work. In fact, if you allow conda to choose the versions for you when creating the environment, these will not match sandpyper requirments when imported with pip and you will get the conflicts you noted in issue number 7.
5. see point number 3.
6. absolutely, thanks for noting this!

The init typos and the pysal dependency errors are both due to the fact that your installed version was a very old one 0.0.2 (instead of the new 1.0.0 ).
I suspect that by allowing conda to decide for you which package to download (without specifying the package versions) caused an error somewhere (most probably GDAL via Rasterio) and made pip look for an older compatible version of Sandpyper which happened to be the one that had the above mentioned issues.

So, I now:
- correct the readme
- package everything in conda

Thanks for this!
