### 2021-09-06 npucino/sandpyper#6 @npucino
By the way, @dbuscombe-usgs can you explain me why being strict with packages versions during installation is not future-proof, assuming these packages versions will not be deleted from servers? Thanks!
