### 2021-09-06 2451c6c
added python=3.8 and removedconfusing richdem mention
