### 2021-09-06 #3666 @crvernon
Thanks @dbuscombe-usgs and @npucino it looks like you are working towards a solution.  Feel free to reach out if you have any questions.
