+++ .github/workflows/build.yml
@@ -2,11 +2,11 @@ on:
     push:
         branches:
             - master
-            - v1
+            - conda_only
     pull_request:
         branches:
             - master
-            - v1
+            - conda_only
 
 name: build
 jobs:
