+++ .github/workflows/docs.yml
@@ -28,6 +28,6 @@ jobs:
             - name: PKG-TEST
               run: |
                   python -m unittest discover
-            - run: pip install mkdocs==1.1.2 mkdocs-material mkdocstrings mkdocs-git-revision-date-plugin mkdocs-jupyter livereload python-markdown-math
+            - run: conda install mkdocs==1.1.2 mkdocs-material mkdocstrings mkdocs-git-revision-date-plugin mkdocs-jupyter livereload python-markdown-math
 
             - run: mkdocs gh-deploy --force
