### 2021-09-07 npucino/sandpyper#6 @dbuscombe-usgs
Hi @npucino, thanks, that sounds like an excellent plan! And yes I believe you are correct about the reason for the installation of the old sandpyper package. I will wait for your changes before attempting again with a fresh installation.

I mentioned about the version numbers because conda probably wants to decide for itself about the versions that are compatible. As bugs are being fixed constantly, my personal philosophy would be to try to use the latest stable versions where possible. Additionally, I think if you want to be restrictive with package versions, you may want to specify a specific version for all packages, not just most ...
