### 2021-09-12 npucino/sandpyper#6 @chrisleaman
Hi @dbuscombe-usgs & @npucino - aren't python packaging issues just fun to deal with 😑😅? Just to add another data point, I did manage to follow the instructions in the "Getting Started" section in the README, get the environment installed and the jupyter notebooks running (on Windows 10 x64).

Still, I would recommend going with the approach @dbuscombe-usgs suggested https://github.com/npucino/sandpyper/issues/6#issuecomment-912860090 and creating a `enviroment.yml` file. This would simplify installation into one command (`conda env create --file sandpyper.yml`) and be less prone to conflicts if users want to install additional packages in the same environment to do their own analysis. In future, it might also be worth putting sandpyper on conda-forge as well.
