### 2021-09-12 #3666 @chrisleaman
> 👋 @chrisleaman, please update us on how your review is going (this is an automated reminder).

Sorted out package installation issues, working through example notebooks at the moment.
