### 2021-09-12 npucino/sandpyper#7 @chrisleaman
[JOSS REVIEW] Notebook #2 incorrect label_k values?
Comments are for https://github.com/openjournals/joss-reviews/issues/3666#issuecomment-913647064.

Following `2 - Profiles extraction, unsupervised sand labelling and cleaning.ipynb` I tried plotting the given `label_k` values for `leo_20180606`, with the profile dataframe I saved but they don't look quite right (see plot below).  Can you confirm the values given in `water_dict`, `no_sand_dict` etc in the notebook are correct? Or is the purpose of the `label_correction.gpkg` to fix this? I'm also wondering if there is some random state that results in different label numbers if you rerun the kmeans clustering?

https://github.com/npucino/sandpyper/blob/ce542c6f8a8f7a046e7d0017b29dfacc0420cbe0/examples/2%20-%20Profiles%20extraction%2C%20unsupervised%20sand%20labelling%20and%20cleaning.ipynb#L2836-L2839

![WQyEhjdWsb](https://user-images.githubusercontent.com/7591552/132973763-3c722b99-0cdd-416e-869f-1d2eeed88f75.png)
