### 2021-09-12 npucino/sandpyper#8 @chrisleaman
[JOSS REVIEW] Cannot run P.cleanit()
Comments are for https://github.com/openjournals/joss-reviews/issues/3666#issuecomment-913647064.

Following `2 - Profiles extraction, unsupervised sand labelling and cleaning.ipynb`, I get a `MergeError: Merge keys are not unique in right dataset; not a one-to-one merge` error when running `P.cleanit`.
https://github.com/npucino/sandpyper/blob/ce542c6f8a8f7a046e7d0017b29dfacc0420cbe0/examples/2%20-%20Profiles%20extraction%2C%20unsupervised%20sand%20labelling%20and%20cleaning.ipynb#L3396-L3399

Looks like the validation when merging the dataframe is failing, potentially because of multiple matches in the dataframe?
https://github.com/npucino/sandpyper/blob/ce542c6f8a8f7a046e7d0017b29dfacc0420cbe0/sandpyper/common.py#L2268-L2269

I saved the `to_clean_classified` and `to_update_finetune` dataframes before the error is thrown:
- [to_clean_classified.csv](https://github.com/npucino/sandpyper/files/7149115/to_clean_classified.csv)
- [to_update_finetune.csv](https://github.com/npucino/sandpyper/files/7149119/to_update_finetune.csv)
