### 2021-09-12 npucino/sandpyper#9 @chrisleaman
[JOSS REVIEW] Example notebook minor suggestions
Hi @npucino,

Great work in creating these example notebooks, it was pretty easy for me to work through all three notebooks and understand how everything in the package works.

My comments below are mainly minor suggestions about providing some additional explaination for anyone using these notebooks. Comments are for https://github.com/openjournals/joss-reviews/issues/3666#issuecomment-913647064. I've created two other issues for some trickier problems (https://github.com/npucino/sandpyper/issues/7, https://github.com/npucino/sandpyper/issues/8).

## 1 - Introduction and data preparation.ipynb

1. Can you define the term "LoD" before you use it here? I had to scroll down a bit before I got the definition.
https://github.com/npucino/sandpyper/blob/ce542c6f8a8f7a046e7d0017b29dfacc0420cbe0/examples/1%20-%20Introduction%20and%20data%20preparation.ipynb#L103

2. When talking about the demo data, I think it should be more clear that these are the inputs that sandpyper needs (and not the output that sandpyper generates). It would also be helpful to briefly describe how each input is typically created, i.e. orthos and dems are stitched together using photogrammetry software (Pix4D), while transect and mask files are manually created in a GIS software (QGis, ARCGIS). You've provided examples later in the notebook, it'd be good just add a summary sentence or two here as well.
https://github.com/npucino/sandpyper/blob/ce542c6f8a8f7a046e7d0017b29dfacc0420cbe0/examples/1%20-%20Introduction%20and%20data%20preparation.ipynb#L134

3. We need to change the path of a lot of path varibles, depending where we've saved the test_data. I suggest defining the parent path first and constructing the other paths based on this. Something like:

```
from pathlib import Path

test_data_folder = 'C:\Users\Chris\Desktop\sandpyper\examples\test_data'
ortho_path=Path(test_data_folder,"\orthos_1m\leo_20180920_ortho_resampled_1m.tif")
watermasks_path=Path(test_data_folder,"\clean\watermasks.gpkg")
shoremasks_path=Path(test_data_folder,"\clean\shoremasks.gpkg")
label_corr_path=Path(test_data_folder,"\clean\label_corrections.gpkg")
transect_path=Path(test_data_folder,"\transects\leo_transects.gpkg")
transect_lod_path=Path(test_data_folder,"\lod_transects\leo_lod_transects.gpkg")
```

https://github.com/npucino/sandpyper/blob/ce542c6f8a8f7a046e7d0017b29dfacc0420cbe0/examples/1%20-%20Introduction%20and%20data%20preparation.ipynb#L1164-L1169

4. Can you double check the leo features in the label corrections file (`examples/test_data/clean/label_corrections.gpkg`) are in the correct location? Looks like some CRS mistake when I try load it in QGIS. The mar features in the correct location though.
https://github.com/npucino/sandpyper/blob/ce542c6f8a8f7a046e7d0017b29dfacc0420cbe0/examples/1%20-%20Introduction%20and%20data%20preparation.ipynb#L1418

5. I think an additional sentence explaining why points are being classified here would be good. Also update the reference to example notebook 2.
https://github.com/npucino/sandpyper/blob/ce542c6f8a8f7a046e7d0017b29dfacc0420cbe0/examples/1%20-%20Introduction%20and%20data%20preparation.ipynb#L1431

## 2 - Profiles extraction, unsupervised sand labelling and cleaning.ipynb

1. Again, easier to use the `pathlib.Path` approach here to define the parent folder rather than changing each variable.
https://github.com/npucino/sandpyper/blob/ce542c6f8a8f7a046e7d0017b29dfacc0420cbe0/examples/2%20-%20Profiles%20extraction%2C%20unsupervised%20sand%20labelling%20and%20cleaning.ipynb#L118-L120

2. Add the command to export the dataframe to csv: `P.profiles.to_csv('profiles.csv')`
https://github.com/npucino/sandpyper/blob/ce542c6f8a8f7a046e7d0017b29dfacc0420cbe0/examples/2%20-%20Profiles%20extraction%2C%20unsupervised%20sand%20labelling%20and%20cleaning.ipynb#L2799

## 3 - Profile dynamics.ipynb
(I couldn't link to line numbers in this file - perhaps the notebook is too big for github?)

1. **Section: "Plot: LoD normality check"**: The normality check plots show that the distributions fail the normality checks. Can you talk about what (if any) implications this has?

2. **Section: "Plot: multi-scale MECS and volumetrics"**: The plots for `D.plot_transect_mecs(location='mar',tr_id=10)` and `D.plot_transect_mecs(location='mar', lod=D.lod_df, tr_id=10)` seem fairly similar (though I see some variation in the shaded area). Could you add a couple of sentences discussing the difference between using LoD and not?
