### 2021-09-12 #3666 @dbuscombe-usgs
I will try again to install and run the tests / notebooks, so I can complete the 'functionality' section of the checklist, otherwise I have read through the paper and documentation and they are both good as-is
