### 2021-09-13 #3666 @npucino
Thx @chrisleaman and @dbuscombe-usgs for the reviews.
I am currently working on:
1. investigate the bugs reported by @chrisleaman
2. move the package into conda-forge as a conda recipe

So, maybe @dbuscombe-usgs you can wait until I fix those issues before running trough the code. Just a suggestion.

CHeers!
