### 2021-09-13 npucino/sandpyper#7 @npucino
Hi @chrisleaman , I figured this out.
Basically, this cell causes the issue:
```
# Based on our observations on a dataset comprising 87 surveys, 10 clusters (k=10) is generally a good tradeoff.

opt_k={'leo_20180606': 10,
 'leo_20180713': 10,
 'leo_20180920': 10,
 'leo_20190211': 10,
 'leo_20190328': 10,
 'leo_20190731': 10,
 'mar_20180601': 10,
 'mar_20180621': 10,
 'mar_20180727': 10,
 'mar_20180925': 10,
 'mar_20181113': 10,
 'mar_20181211': 10,
 'mar_20190205': 10,
 'mar_20190313': 10,
 'mar_20190516': 10}
```
The above cell, I will render it in Markdown rather than code so that other people don't fall in this trap.

Basically, when following the notebook, running
```opt_k=get_opt_k(sil_df, sigma=0 )```, computes a sub-optimal number of clusters **to use for KMeans**.
These values are what it should be used.

I added the above cell just to show how one could customise the number of clusters in each survey to use, and I made the point that 10 clusters is a good number of clusters based on our experience, that's why surveys are set to 10.

I replicated your image by running that cell, therefore using k = 10 in all surveys, and by using rule-based symbology in Qgis for survey in St. Leonards 20180606 ( note that when using rule-based symbology, if a rule is not satisfied it doesn't break the classification just no points get rendered). Using the opt_k function, for that survey **the suboptimal k is 11.**

In fact, I think your images doesn't actually have label_k = 10, you should have only 10 labels starting from 0 so the max k should be 9. **Can you double check please?**

If you don't run that cell and use k=11 for that survey, this is what you get:
![image](https://user-images.githubusercontent.com/44853536/133020283-fc0e17fb-fc0a-42f9-b13d-f7af6e047aaa.png)

Then, you will be able to actually run` P.cleanit()` (fixing the other bug), and obtain this:

![image](https://user-images.githubusercontent.com/44853536/133020382-f5079a24-ac44-4266-bd91-ede8b8efa94b.png)

**Can you please confirm that was the issue by just skipping that misleading code cell??**
Thanks!
