### 2021-09-13 npucino/sandpyper#8 @npucino
Thx @chrisleaman for reporting this.
By following what I wrote [here ](https://github.com/npucino/sandpyper/issues/7#issuecomment-917811812) you should be able to make it work.

But, I need to figur out this error and why I get duplicates but I think it is related to the fact that some labels_k are present in the polygons used to correct the data but not in the actual profile data.
