### 2021-09-13 npucino/sandpyper#9 @npucino
Thanks @chrisleaman for this review! Very useful and constructive suggestions that I will implement very soon.

For now I want to just reply this:

`Can you double check the leo features in the label corrections file (examples/test_data/clean/label_corrections.gpkg) are in the correct location? Looks like some CRS mistake when I try load it in QGIS. The mar features in the correct location though.`

Unfortunately, storing all locations (multiple locations and multiple CRSs) into one single Geopandas.Geodataframe allow only 1 CRS to use. Thus, I digitised all correction polygons with EPSG 32754, which is the correct one for Marengo but the wrong one for St. Leonards (EPSG 32755). Unfortunately,  we would need either:
1. one geodataframe per location with the actual location CRS
2. a geodataframe which stores multiple CRSs, say, some rows with EPSG 32754 and others EPSG 32755, which is not possible to my knowledge.

I must make clear this somewhere, very important, thanks for reporting this!
