### 2021-09-15 npucino/sandpyper#7 @npucino
Hi @chrisleaman , I checked your profile-cleaned file and I see the confusion now!

The `label_k `field is **not updated,** rather, `P.cleanit()` adds a column named `pt_class`.
That column holds the final point classes, cleaned, in this case water, sand, veg, no_sand.

here is the pt_class column displayed of your profile-cleaned.csv file!

![image](https://user-images.githubusercontent.com/44853536/133351520-c3017ac7-5fa5-4e58-a738-e8a181a1987d.png)
