### 2021-09-15 npucino/sandpyper#7 @chrisleaman
Thanks @npucino, everything makes sense now - I didn't realise that extra column had been added! I think an extra sentence or two in the notebook just stating the additional `pt_class` column is created would be very helpful.

Re: performance issues, if you haven't already, I'd recommend looking into using https://github.com/pyutils/line_profiler to see exactly where your code is slow. It identifies which lines in your code are taking the most time, so you can focus on increasing performance just on those lines. Sometimes the results are surprising and you can get some easy wins using this!

Feel free to close this issue when you're ready 😊
