### 2021-09-15 npucino/sandpyper#7 @npucino
Thanks for the suggestion @chrisleaman , forked now!
This afternoon I will take care of the rest of the issues as well.

Cheers!
