### 2021-09-17 174df4b
added LoD vs no LoD explaination and mod some headings
