### 2021-09-20 npucino/sandpyper#8 @npucino
Ok I found out why some points got duplicated when the number of k was 10 for all the surveys to be clustered with KMeans, it was a tricky one but luckily you spot it.
Basically it was an **overlap issue with label correction polygons.** Let me explain you.

Label correction polygons can overlap because they are only operating on a survey basis (raw date and location) and only target those points within their boundaries that have the `label_k `attribute equals to the label correction polygon attribute `target_label_k`.
However, if polygons targeting the same `label_k `overlap and in their overlap area they include points with their `target_label_k`, the new class **MUST BE THE SAME**, otherwise it doesn't make sense as nobody knows that point in that shared area what new class should take!

What happened was that I digitised label correction polygons by looking at their `labels_k `as per returned by KMeans using the correct number of k ( the opt_k dictionary). My polygons overlapped of course because no overlapping areas contained points with label_k. But when running the wrong number of k, the label_k of th epoints changed and the polygons suddently became irrelevant! Then, in one survey (leo_20190731), there was one point with `label_k=6` which happened to be right in a overlap area of two label correction polygons with `target_label_k=6 `but **with 2 different new classes attribute** (_sand_ and _veg_) and this caused the duplication of these points and the non-uniqueness of their IDs and the impossibility to merge etc etc until the Error.
See below image, that point caused the error as it is sitting right in the overlap of the yellow and purple polygons.

![image](https://user-images.githubusercontent.com/44853536/133947223-a9df07ef-1147-4b8f-9d4b-1452707989c7.png)

Of course, this wouldn't happen if the user do not rerun KMeans with different random seed or different number of k per survey!
I hope I explained it well.

Here is what I did to correct this error.
Instead of simply write a warning and relying on the users to not to overlap label correction polygons targeting the same `label_k `but assigning different new classes, I decided to write a function that check for inconsistencies after the polygon creation, together with clearly stating this rule in the DOC. This would allow users to be more flexible because when the number of points is large (if for example you sample a point every 10cm along a transect), points can get dense and is quicker to just overlap.

Here is the function that takes care of that and checks the aforementioned rule:
```python
def check_overlaps_poly_label(label_corrections, profiles):
    for loc in label_corrections.location.unique():
        print(loc)
        for raw_date in label_corrections.query(f"location=='{loc}'").raw_date.unique():
            print(f"Check raw_date {raw_date}")
            for target_label_k in label_corrections.query(f"raw_date=={raw_date}").target_label_k.unique():
                print(f"Check target_label_k {target_label_k}")

                date_labelk_subset=label_corrections.query(f"raw_date=={raw_date} and target_label_k=={int(target_label_k)}")

                # if more than one polygons target the same label k, check if they overlap
                if len(date_labelk_subset)>1:
                    print(f"There are {len(date_labelk_subset)} corr polys")

                    # check if there are any one of them that overlaps
                    for i,z in comb(range(len(date_labelk_subset)),2):
                        intersection_gdf = overlay(date_labelk_subset.iloc[[i]], date_labelk_subset.iloc[[z]], how='intersection')

                        if not intersection_gdf.empty:

                           # check if the overlapping polygons have assigns different new_classes
                            if any(intersection_gdf.new_class_1 != intersection_gdf.new_class_2):

                                # if overlap areas assign different classes, check if this area contains points.
                                # if contains points, raise an error as it does not make sense and the polygons must be corrected
                                # by the user

                                pts=profiles.query(f"location=='{loc}' and raw_date=={raw_date}")
                                pts['coordinates']=pts.coordinates.apply(coords_to_points)
                                pts_gdf=gpd.GeoDataFrame(pts, geometry='coordinates', crs=crs_dict_string[loc])

                                fully_contains = [intersection_gdf.geometry.contains(mask_geom)[0] for mask_geom in pts_gdf.geometry]

                                if True in fully_contains:
                                    idx_true=[i for i, x in enumerate(fully_contains) if x]
                                    raise ValueError(f"There are {len(intersection_gdf)} overlapping polygons in {loc} raw_date {raw_date} targeting label_k = {target_label_k} that assign to the same points different the classes {intersection_gdf.loc[:,'new_class_1'][0], intersection_gdf.loc[:,'new_class_2'][0]}. This doesn't make sense, please correct your label correction polygons. You can have overlapping polygons which act on the same target label k, but in this case they MUST assign the same new class to these points.")

    print("Check terminated successfully")
```

Now users can create label correction polygons which target the same label k without stressing too much on being super precise and not overlapping. If an issue of overlapping area of 2 polygons which target the same label k but assign different class should contain a point with such target k, this function will return a `ValueError `and point to what location, what date and what target_label_k created the error, allowing the user to quickly return to the GIS and correct the polygons. The message is like this one:

```python
ValueError: There are 1 overlapping polygons in leo raw_date 20190731 targeting label_k = 6 that assign to the same points different the classes ('sand', 'veg'). This doesn't make sense, please correct your label correction polygons. You can have overlapping polygons which act on the same target label k, but in this case they MUST assign the same new class to these points.
```

I am now updating the doc with this warning, then I will place this check directly in the **cleanit method.**
