### 2021-09-20 12e5965
added label corr polygon overlap check, updated DOCs and NBs
