+++ sandpyper/common.py
@@ -15,11 +15,13 @@ from shutil import move, copy
 import numpy as np
 import pandas as pd
 import geopandas as gpd
+from geopandas.tools import overlay
 import seaborn as sb
 from tqdm.notebook import tqdm_notebook as tqdm
 from fuzzywuzzy import fuzz
 import itertools
 from itertools import product, groupby
+from itertools import combinations as comb
 
 import math
 from math import tan, radians, sqrt
@@ -2210,7 +2212,7 @@ def classify_labelk(labelled_dataset,l_dicts, cluster_field='label_k', fill_clas
 def cleanit(to_clean, l_dicts, cluster_field='label_k', fill_class='sand',
             watermasks_path=None, water_label='water',
             shoremasks_path=None, label_corrections_path=None,
-            default_crs={'init': 'epsg:32754'}, crs_dict_string=None,
+            crs_dict_string=None,
            geometry_field='coordinates'):
 
     print("Reclassifying dataset with the provided dictionaries." )
@@ -2228,6 +2230,13 @@ def cleanit(to_clean, l_dicts, cluster_field='label_k', fill_class='sand',
         if os.path.isfile(label_corrections_path):
             label_corrections=gpd.read_file(label_corrections_path)
             print(f"Label corrections provided in CRS: {label_corrections.crs}")
+
+            # validate label correction polygons
+            if crs_dict_string != None:
+                check_overlaps_poly_label(label_corrections,to_clean,crs_dict_string)
+            else:
+                check_overlaps_poly_label(label_corrections,to_clean,label_corrections.crs.to_epsg())
+
             processes.append("polygon finetuning")
             to_update_finetune=pd.DataFrame()
 
@@ -5883,3 +5892,51 @@ def LISA_site_level(
             lisa_df = pd.concat([lisa_df, gdf_input], ignore_index=True)
 
     return lisa_df
+
+
+def check_overlaps_poly_label(label_corrections, profiles,crs):
+    """
+    Function to check wether overlapping areas of label correction polygons targeting the same label_k in the same surveys but assigning different new classes do not contain points that would be affected by those polygons.
+    Args:
+        label_corrections (gpd.GeoDataFrame): GeodataFrame of the label correction polygons.
+        profiles (gpd.GeoDataFrame): Geodataframe of the extracted elevation profiles.
+        crs (dict, int): Either an EPSG code (int) or a dictionary. If dictionary, it must store location codes as keys and crs information as values, in dictionary form (example: {'init' :'epsg:4326'}).
+
+    """
+    for loc in label_corrections.location.unique():
+        for raw_date in label_corrections.query(f"location=='{loc}'").raw_date.unique():
+            for target_label_k in label_corrections.query(f"raw_date=={raw_date}").target_label_k.unique():
+
+                date_labelk_subset=label_corrections.query(f"raw_date=={raw_date} and target_label_k=={int(target_label_k)}")
+
+                # if more than one polygons target the same label k, check if they overlap
+                if len(date_labelk_subset)>1:
+
+                    # check if there are any one of them that overlaps
+                    for i,z in comb(range(len(date_labelk_subset)),2):
+                        intersection_gdf = overlay(date_labelk_subset.iloc[[i]], date_labelk_subset.iloc[[z]], how='intersection')
+
+                        if not intersection_gdf.empty:
+
+                           # check if the overlapping polygons have assigns different new_classes
+                            if any(intersection_gdf.new_class_1 != intersection_gdf.new_class_2):
+
+                                # if overlap areas assign different classes, check if this area contains points.
+                                # if contains points, raise an error as it does not make sense and the polygons must be corrected
+                                # by the user
+
+                                pts=profiles.query(f"location=='{loc}' and raw_date=={raw_date}")
[…]
