### 2021-09-20 47d7708
updated check_overlaps_poly_label fn
