### 2021-09-20 a629022
updated check overlap label corr polygons fn
