### 2021-09-20 aaa9f19
fic check overlap polygons
