### 2021-09-20 1908e2e
added crs_dict_string par to cleanit method in test
