+++ tests/test_sandpyper.py
@@ -257,7 +257,8 @@ class TestProfileSet(unittest.TestCase):
         cls.P.cleanit(l_dicts=l_dicts,
           watermasks_path=watermasks_path,
           shoremasks_path=shoremasks_path,
-          label_corrections_path=label_corrections_path)
+          label_corrections_path=label_corrections_path,
+          crs_dict_string=cls.P.crs_dict_string)
 
         cls.check_watermask_mar_post=cls.P.profiles.query("location == 'mar' and raw_date==20181113 and distance < 10").shape
         cls.check_watermask_leo_post=cls.P.profiles.query("location == 'leo' and raw_date==20180920 and distance < 10").shape
