### 2021-09-20 npucino/sandpyper#8 @chrisleaman
Looks like a tricky error to find! Glad you managed to sort it out - a warning to the user, like you suggested, should work well.
