+++ sandpyper/common.py
@@ -5921,9 +5921,10 @@ def check_overlaps_poly_label(label_corrections, profiles,crs):
                            # check if the overlapping polygons have assigns different new_classes
                             if any(intersection_gdf.new_class_1 != intersection_gdf.new_class_2):
 
-                                # if overlap areas assign different classes, check if this area contains points.
+                                # if overlap areas assign different classes, check if this area contains points with label_k equal to both polygons target_label_k..
                                 # if contains points, raise an error as it does not make sense and the polygons must be corrected
                                 # by the user
+
                                 pts=profiles.query(f"location=='{loc}' and raw_date=={raw_date} and label_k=={int(target_label_k)}")
 
                                 if isinstance(pts.iloc[0]['coordinates'],Point):
