### 2021-09-20 c92b897
test crs_dict_string instead of cls.P.crs_dict_string
