### 2021-09-20 npucino/sandpyper#8 @npucino
Triki but important. Now I am struggling to run the function in the github environment while testing, which works perfectly local. As soon as I find what is causing this error I update you here. Cheers
