### 2021-09-20 62848a0
replaced to_clean with to_clean_classified in check poly labels overlap
