### 2021-09-21 npucino/sandpyper#9 @npucino
Here there are the suggestions implemented.
>**Sorry for the rudimental images snapshots,** I don't know how to reference notebook lines as links and couldn't find a solution in 5 minutes of googling.

### 1.1 Earlier definition of LoD
I added its definition straight directly in the introduction.

![image](https://user-images.githubusercontent.com/44853536/134099413-c42c14cf-6291-4beb-a122-4516a5d07dc8.png)

### 1.2 Better definition of input data
I added a better explaination of each input, making clear that these are inputs and not what Sandpyper produces.

![image](https://user-images.githubusercontent.com/44853536/134099443-f2140252-b143-4f9e-a478-7b4fd75bd04d.png)

### 1.3 & 2.1 Implementation of pathlib Path to manage paths
I used the strategy you suggested and it is way better.

![image](https://user-images.githubusercontent.com/44853536/134099524-7b8bb8e9-3224-4592-834e-0e8eef5ccbbb.png)

### 1.4 CRS misplacement
Explained here: https://github.com/npucino/sandpyper/issues/9#issuecomment-917816519

### 1.5 additional explanation of point classification importance and updated link to notebook 2
I wrote an extra paragraph where I introduce the challenges and the why I coded this way (for now).
And updated the link to NB-2.

![image](https://user-images.githubusercontent.com/44853536/134099671-2d43a325-d089-4829-9f52-dd2ab68cb6b0.png)

### 2.2 save to .csv command
Added.
![image](https://user-images.githubusercontent.com/44853536/134099260-c0a3e79f-9db2-424b-a6de-6b31d8b964d5.png)

### 3.1 Normality check importance and consequences explanation
I better explained why it is important to test the normality of the error distributions and the consequences on the decision of most suitable error metric, with reference to one of the most important literature that best explain this concept for this type of data.

![image](https://user-images.githubusercontent.com/44853536/134099874-a24c9e29-968e-452d-868e-aa24aef482a9.png)

### 3.2 Plot transect mecs with and without LoD better explanation
I added a few words on the importance of using LoD despite the difference being small.

![image](https://user-images.githubusercontent.com/44853536/134099781-d9d159db-ac26-4fd3-87f2-19591e283ba3.png)
![image](https://user-images.githubusercontent.com/44853536/134099805-a48b0d20-2f8e-47fe-aee2-ea3c6df2c937.png)

Cheers
