### 2021-09-21 ca5641d
updated roadmap and doc removed autom poly label check
