### 2021-09-21 npucino/sandpyper#8 @npucino
If you agree a warning is enough (for now), I added a big note in the DOC, in the label correction file section of the data cleaning chapter.

![image](https://user-images.githubusercontent.com/44853536/134103142-589a00a3-9f04-4b91-8471-16d4770db59c.png)

However, the function I created works perfectly in my local Jupyter notebook setup, but for some reasons it doesn't work when running in the GitHub VM during testing the package.
Therefore I decided to stop trying to make it work, adopt your suggestion of the warning, but I __added it as a roadmap milestone__, because even if it might not be crucial at the moment, I think an automatic check is beneficial.

So I will take the time to make this check work in the near future.
