### 2021-09-21 2c78ee4
restored cls.P.crs_dict_string in cleanit
