### 2021-09-21 b3cda2c
test cls.P.crs_dicts
