### 2021-09-21 6cf50d9
test Path method in testing script
