### 2021-09-22 62aedc2
correcter cls.P.profiles.shape print
