### 2021-09-22 1041f67
added np.random seed 10
