### 2021-09-22 npucino/sandpyper#9 @chrisleaman
Hi @npucino, great work with these changes! I really like the additional guidance and discussion you've provided, especially in notebook 3. It'll be especially helpful for those that aren't familiar with these LoD concepts (like me!).

This issue was addressed with the following commits:

- 785ffca04c52c877002e390571d669954a6442cf
- 174df4bfebb0a5bdf9713fc482072c29799faf04
- 7b2cf21418283b5d1108a5dd07016fdbee44d834
- 324f2c96de90fd6611a7775ee87d82a6fa9905f7
- 5c8a4549c435c7d1ad2a5daf1aba1b480599a4cf
- 785ffca04c52c877002e390571d669954a6442cf
- 12e5965ea163f8858765574c01759e8f1594da42
- ab8411ee195e6756c9bcf83c088878c9783f8d88
