### 2021-09-22 npucino/sandpyper#8 @chrisleaman
Hi @npucino, I agree with your approach - a warning is fine for now if you're struggling to get it to work on Github 👍
