### 2021-09-22 #3666 @chrisleaman
@crvernon - Just an update, I'm happy with code and documentation as @npucino has addressed the issues I've raised. I have a couple of minor suggestions for the manuscript which I will finalise and create an issue for by the end of the week.
