### 2021-09-22 npucino/sandpyper#8 @npucino
Hi thanks @chrisleaman . I see that you fetched the commits related to the issues, I am sorry I didn't do it myself.
Now I was trying to see the updated part in the DOC but it seems that GitHub action never rendered the new version. I look into it now!
