### 2021-09-22 41475d5
commented out ubuntu build check for DOC rendering
