### 2021-09-22 019eb17
restoring strict packages requirments
