### 2021-09-22 869fd79
use pip for mkdocs etc
