### 2021-09-22 d6e8952
Bump version: 1.1.0 → 1.2.0
