### 2021-09-22 1330ec9
updated versions and readme prior conda test
