### 2021-09-22 a0c6565
eliminated useless testing files
