### 2021-09-22 #3666 @crvernon
:mega:  Mid-week rally!  Just checking in to see how things are going.

:open_hands: Thanks @chrisleaman for the notification!  Glad to see things are progressing well.

@dbuscombe-usgs and @npucino it looks like you both are still working out some bugs so keep up the good work!

Have a great and productive rest of the week!
