### 2021-09-22 #3666 @npucino
Hi @crvernon , I packaged the code in conda and I already pull request its recipe to the team:
https://github.com/conda-forge/staged-recipes/pull/16228

I am just waiting to be added to the conda server and then test with @dbuscombe-usgs if everything works well with this strategy!
I do not know how long this usually take, hopefully not too much.

Cheers!
Nick
