### 2021-09-27 #3666 @npucino
I am having an hard time relaxing the requirements for Sandpyper right now.

Basically, as per [conda request](https://github.com/conda-forge/staged-recipes/pull/16228) and @dbuscombe-usgs request I was trying to use the most recent versions of Sandpyper dependencies rather than providing strict versions of those.
Apart from minor changes to the code (CRS definition and some minor Pysal import changes), the code should run smoothly with no issues.
Note that the tests and Jupyter Notebooks that I created are based on the data I generated with Sandpyper within the environment with strict dependencies requirements.

### Issue
When I create a fresh new environment with no dependencies versions specified except for Python = 3.8, I have (for now) one issue that I cannot understand where it comes from.

In the unsupervised labelling procedure, within my [get_sil_location](https://github.com/npucino/sandpyper/blob/a0c65657c2175c9bbd7a7b084affc30859385521/sandpyper/common.py#L1915), I use Scikit-Learn [KMeans](https://scikit-learn.org/stable/modules/generated/sklearn.cluster.KMeans.html) and [Silhouette score](https://scikit-learn.org/stable/modules/generated/sklearn.metrics.silhouette_score.html?highlight=silhouette#sklearn.metrics.silhouette_score) functions to cluster my points and to compute silhouette scores. I set an integer as random state so that I obtain always the same results.
The expected results must be obtained in order to run the tests and to follow along the notebooks I provided to see how Sandpyper works.
The code works, but with the relaxed environment the __labels KMeans assign are slightly different__, which affects the silhouette scores, the suboptimal number of clusters and eventually the labels the points will be assigned to.
In order to run tests and provide users with actionable Jupyter Notebooks, these little changes would require me to go through the dataset again, create new polygons of corrections and re-identify the classes of all the test surveys, which takes quite a bit of time and I ideally would like no to go through that route at this stage if possible.
So I looked at the [changelog](https://scikit-learn.org/stable/whats_new/v1.0.html) of sklearn (changes from 0,24 to 1.0 versions) and __some minor changes occured in the KMeans algorithm__, but those seems to be related to performances rather than random state.
I also tested environments completely relaxed except for `sklearn==0.24.1` which is the one I used for the test data, but the inconsistencies are still there.

My main suspect was the random state but it looks like that is not the case.
I don't understand what is causing to change the output from the `get_sil_location` function, despite it is working as expected.
If I use the strict environment, everything works as expected.

### Question
As @chrisleaman already proved, Sandpyper is working as expected and passing tests as is (hosted in PyPI with strict dependencies requirements).
Is hosting Sandpyper in Anaconda a must do for this package to be published in JOSS?
It is of course very important to relax the requirements and upload the package to conda and I really aim to do it in the near future.
But if this is not super strictly necessary __in the context of publishing in JOSS__, I would rather go through the process in a later stage.

Cheers!
Nick
