### 2021-09-28 #3666 @chrisleaman
Re https://github.com/openjournals/joss-reviews/issues/3666#issuecomment-927454043, I'm fine for this JOSS review if `sandpyper` is not on conda.

We should still confirm that @dbuscombe-usgs is able to get it installed and running in his environment or if there are still other issues.
