### 2021-09-28 npucino/sandpyper#13 @npucino
Hi @chrisleaman , your suggestions greatly improved my manuscript, thanks a lot.

I implemented all your suggestions except for:

9. `Line 28: "monitoring program mobilies more than":` Should this be "monitoring program has mobilized more than".<br>
The program is still active and new datasets keep flowing in. I realised my wording misled you as I wrote that sentence using the past tense, so I now modified it into the present tense to make sure it is clear citizen scientists are still flying. And also replaced `mobilies `to `mobilises`.
Here is the snippet:
`For instance in Victoria (Australia), a citizen-science UAV-SfM monitoring program mobilises more than 150 volunteers to fly UAVs on 15 sites every six weeks during the last three years.`

6. `Line 23: "at the mesoscale, a spatiotemporal resolution":` Should this be "at the mesoscale, and at a spatial temporal resolution"?<br>
I used this wording instead: `at the mesoscale, which is an appropriate spatiotemporal scale for coastal managers to work with (Thom et al., 2018).

> here is the reference if interested: [Thom, B. G. et al. National sediment compartment framework for Australian coastal management. Ocean Coast. Manag. 154, 103–120 (2018)](https://linkinghub.elsevier.com/retrieve/pii/S0964569117306129)

Thanks again for this review!
Cheers
`
