### 2021-09-28 abb191e
implemented chrislehman JOSS review suggestions
