+++ paper/paper.md
@@ -29,23 +29,23 @@ bibliography: paper.bib
 
 **Sandpyper** is a Python package that automates profile-based volumetric and altimetric sandy beaches analysis from a large amount of digital surface models and orthophotos.  It includes functionalities to facilitate the cleaning of the elevation data from unwanted non-sand points or swash areas (where waves run up on the beach slope and 3D reconstruction is inaccurate) and to model beachface behavioural regimes using the Beachface Cluster Dynamics indices.
 
-# Intro
+# Introduction
 
-Coastal zones host 40% of the world population [@Martinez2007-ux]  and it is increasing, especially in least developed countries [@Neumann2015-ph]. Sandy beaches, amongst other ecoservices [@Barbier2011-qp], protect inland assets from coastal erosion, dissipating stormy waves energy on their shores. Mitigating beach erosion typically involves the establishment of topographic monitoring programs in key locations (erosional hotspots) to quantify beach dynamics, erosion/deposition volumes, recovery times and model coastal resilience or risk to erosion. High temporal and spatial resolution topographic data is ideal, but expensive with most of the ordinary beach surveying methods.
-Unmanned Aerial Vehicles (UAVs) and Structure from Motion algorithms (UAV-SfM) are emerging as the best platform and methodology to obtain cost-effective high-quality beach topographic data (as Digital Surface Models, DSMs) [@Goncalves2015-qr] at the mesoscale, a spatiotemporal resolution appropriate for coastal management [@Thom2018-kf]. Consequently, researchers already use UAV-SfM to monitor beach dynamics around the world, but it has been limited so far to a few sites and a few multitemporal dates. However, UAV-SfM technology is mature and reliable enough to allow wider-scale and longer-term monitoring projects.
-For instance in Victoria (Australia), a citizen-science UAV-SfM monitoring program mobilises more than 150 volunteers to fly UAVs on 15 sites every six weeks for three years. To date, volunteers flew 350 times, enabling the creation of a DSM and an orthophoto per survey (uncompressed file sizes from 5-100 Gb each), which generates an unprecedented archive of imagery which can be reliably used to monitor high-frequency sandy beach volumetric dynamics and behaviors [@Pucino2021-ox].
+Coastal zones host 40% of the world population [@Martinez2007-ux]  with continued expected growth to be focused in least developed countries [@Neumann2015-ph]. Sandy beaches, amongst other ecoservices [@Barbier2011-qp], protect inland assets from coastal erosion, dissipating stormy waves energy on their shores. Mitigating beach erosion typically involves the establishment of topographic monitoring programs in key locations (erosional hotspots) to quantify beach dynamics, erosion/deposition volumes, recovery times and model coastal resilience or risk to erosion [@Turner2016-dv, @Ruessink2019-yl]. High temporal and spatial resolution topographic data is ideal, but expensive with most of the ordinary beach surveying methods.
+Unoccupied Aerial Vehicles (UAVs) and Structure from Motion algorithms (UAV-SfM) are emerging as ideal platform and methodology to obtain cost-effective high-quality beach topographic data (as Digital Surface Models, DSMs) [@Goncalves2015-qr] at the mesoscale, which is an appropriate spatiotemporal scale for coastal managers to work with [@Thom2018-kf]. Consequently, researchers already use UAV-SfM to monitor beach dynamics around the world [@Brunier2016-hq, @Jaud2019-rp], but it has been limited so far to a few sites and a few multitemporal dates. However more recently, UAV-SfM technology is being increasingly used for wider-scale and longer-term monitoring projects.
[…]
