+++ paper/paper.bib
@@ -399,10 +399,10 @@
 }
 
 
-@article{Beuzen2019,
+@ARTICLE{Beuzen2019,
   doi = {10.21105/joss.01890},
   url = {https://doi.org/10.21105/joss.01890},
-  year = {2019},
+  year = 2019,
   publisher = {The Open Journal},
   volume = {4},
   number = {44},
@@ -410,4 +410,211 @@
   author = {Tomas Beuzen},
   title = {pybeach: A Python package for extracting the location of dune toes on beach profile transects},
   journal = {Journal of Open Source Software}
+
+ @ARTICLE{Turner2016-dv,
+    doi = {10.1038/sdata.2016.24},
+    url = {https://www.nature.com/articles/sdata201624},
+    title     = "A multi-decade dataset of monthly beach profile surveys and
+                 inshore wave forcing at Narrabeen, Australia",
+    author    = "Turner, Ian L and Harley, Mitchell D and Short, Andrew D and
+                 Simmons, Joshua A and Bracs, Melissa A and Phillips, Matthew S
+                 and Splinter, Kristen D",
+    abstract  = "Long-term observational datasets that record and quantify
+                 variability, changes and trends in beach morphology at sandy
+                 coastlines together with the accompanying wave climate are rare.
+                 A monthly beach profile survey program commenced in April 1976
+                 at Narrabeen located on Sydney's Northern Beaches in southeast
+                 Australia is one of just a handful of sites worldwide where
+                 on-going and uninterrupted beach monitoring now spans multiple
+                 decades. With the Narrabeen survey program reaching its 40-year
+                 milestone in April 2016, it is timely that free and unrestricted
+                 use of these data be facilitated to support the next advances in
+                 beach erosion-recovery modelling. The archived dataset detailed
+                 here includes the monthly subaerial profiles, available
+                 bathymetry for each survey transect extending seawards to 20 m
+                 water depth, and time-series of ocean astronomical tide and
+                 inshore wave forcing at 10 m water depths, the latter
+                 corresponding to the location of individual survey transects. In
+                 addition, on-going access to the results of the continuing
+                 monthly survey program is described. Machine-accessible metadata
+                 file describing the reported data (ISA-Tab format)",
+    journal   = "Scientific Data",
+    publisher = "Nature Publishing Group",
+    volume    =  3,
+    number    =  1,
+    pages     = "1--13",
+    month     =  apr,
+    year      =  2016,
+    language  = "en"
+  }
+
+
+  @ARTICLE{Ruessink2019-yl,
+    doi = {10.3390/data4020073},
+    title     = "A {Multi-Year} Data Set of {Beach-Foredune} Topography and
+                 Environmental Forcing Conditions at Egmond aan Zee, The
+                 Netherlands",
+    author    = "Ruessink, Gerben and Schwarz, Christian S and Price, Timothy D
+                 and Donker, Jasper J A",
+    abstract  = "Coastal dunes offer numerous functions to society, such as sea
+                 defense and recreation, and host unique habitats with high
+                 biodiversity. Research on coastal dune dynamics has
+                 traditionally focused on the erosional impact of short-duration
+                 (hours to days), high-wave storm events on the most seaward
+                 dune, called the foredune. In contrast, research data on its
+                 subsequent slow (months to years), wind-driven recovery are
+                 rather rare, yet essential to aid studying wind-driven
+                 processes, identifying the most relevant wind-forcing
+                 conditions, and testing and improving dune-growth models. The
+                 present data set contains 39 digital elevation models and 11
+                 orthophotos of the beach-foredune system near Egmond aan Zee,
+                 The Netherlands. The novelty of the data set lies in the
+                 combination of long-term observations (6 years; January 2013 to
+                 January 2019), with high temporal (intervals of 2--4 months) and
[…]
