+++ paper/paper.bib
@@ -409,7 +409,7 @@
   pages = {1890},
   author = {Tomas Beuzen},
   title = {pybeach: A Python package for extracting the location of dune toes on beach profile transects},
-  journal = {Journal of Open Source Software}
+  journal = {Journal of Open Source Software} }
 
  @ARTICLE{Turner2016-dv,
     doi = {10.1038/sdata.2016.24},
@@ -610,11 +610,4 @@
   organization = {QGIS Association},
   year = {2021},
   url = {https://www.qgis.org},
-}
-
-
-
-
-
-
 }
