### 2021-09-28 #3666 @npucino
> Re [#3666 (comment)](https://github.com/openjournals/joss-reviews/issues/3666#issuecomment-927454043),
> [...]
> We should still confirm that @dbuscombe-usgs is able to get it installed and running in his environment or if there are still other issues.

Absolutely. Thanks.
