### 2021-09-28 a0f6dac
test ubuntu py 3.8 and 3.9
