+++ .github/workflows/build.yml
@@ -19,8 +19,8 @@ jobs:
                 config:
                     - { os: windows-latest, py: "3.8" }
                     - { os: macOS-latest, py: "3.8" }
-                    #- { os: ubuntu-latest, py: "3.8" }
-                    #- { os: ubuntu-latest, py: "3.9" }
+                    - { os: ubuntu-latest, py: "3.8" }
+                    - { os: ubuntu-latest, py: "3.9" }
         env:
             SDKROOT: /Library/Developer/CommandLineTools/SDKs/MacOSX.sdk
 
