### 2021-09-29 #3666 @dbuscombe-usgs
I am back from vacation and will be ready to tackle this tomorrow. Apologies for the delay
