### 2021-09-29 #3666 @crvernon
@npucino - to address your earlier question of:

> Is hosting Sandpyper in Anaconda a must do for this package to be published in JOSS?

The answer from my perspective is:  no.  Having this package able to be installed and function given the documentation and claims that you provide is sufficient.  I agree that this would be a nice addition in the future to extend the usability of your package; though it may not be in the current scope of your development.
