### 2021-09-29 npucino/sandpyper#13 @chrisleaman
Great, thanks for clarifying points 9. and 10., happy with what you proposed 👍
