### 2021-09-29 #3666 @dbuscombe-usgs
1. I have removed the previous sandpyper_env conda environment I made before
2. I don't see any instructions for a conda package installation on https://github.com/npucino/sandpyper, so I'm not really sure whether that is yet an option (it seemed like it was, from this thread, but now I'm not so sure). I was expecting the README to be updated with instructions for how you would like to me to install the sanpyper conda-forge package.
3. However, I have now successfully install using the provided instructions on the README (using the suboptimal legacy versions of various packages) and that has given no errors on installation. Progress!
4. The installation process is still really messy - for many of the reasons I previously stated - there are several packages that get installed, and reinstalled. Some get reinstalled several times! Using a mixture of pip and conda to specify a mixture of major, minor, and incremental versions of programs is kinda sloppy, and I refuse to accept this is absolutely necessary, but it WORKS (now), so I'll let it slide!

I agree with @crvernon and @chrisleaman that a conda solution shouldnt be a requirement for JOSS, however I will say that scikit-learn is the major ML library in python, and using a version that is several versions behind the main trunk (we are at 1.0 now) is storing up problems for the future, IMHO, if this is to be maintained for several years. I anticipate you'll be fielding a lot of questions about installation going forward, if the toolbox proves popular (I hope so). However, over time hopefully a solution is provided. I took a quick look at the changelogs and API references for the KMeans functions over time and, as @npucino noted, there doesn't seem to be anything related to random seeds or state, so it is puzzling!

To run the notebooks., the instructions are inadequate for a novice user, who needs to be told they need to download the github package (like I said before, why go to the bother of making a pip package when the user has to download the code anyway?). A few sentences here in the README explaining to the potential user would go a long way

In the first notebook, "1 - Introduction and data preparation", you don't explain where 'demo_data.rar' is located. Its not a huge deal, because I discovered you actually don't need it, but I did waste about 10 minutes looking for it! I looked at the paper dataset, logging into the propeller site as-per the 'data availability' statement, but I dont quickly see how to download that data file. In fact, it is not at all clear how to download any data from the propeller site...

I am able to run all the provided examples. The toolbox is fast, and very useful. The methods work as-described, and the statements made about the software in the paper and supporting info are accurate. Good job!

I'll complete the checklist and APPROVE this for JOSS. I trust the level of documentation for installation and implementation will improve with a growing user base
