### 2021-09-29 24533d3
removed notebook from prereq instruction
