### 2021-09-30 #3666 @npucino
Hi @dbuscombe-usgs , thanks for your feedback!
Let me address your valuable points one at the time.

### Conda-forge package

One of the major request from yourself and conda-forge reviewers is to relax the requirements for Sandpyper.
If I relax the requirements basically telling conda to install the latest versions of packages which are compatible with the user active environment, I get that weird behaviour that I explained in my last comment https://github.com/openjournals/joss-reviews/issues/3666#issuecomment-927454043. This doesn't mean the code doesn't work, it just means that because small differences in the labels assigned by KMeans, I have to re-build the test data from scratch (correction polygons and class dictionaries). This requires some spare time from my side which __I am more than willing to spend__ in my evenings every now and then rather than full time right now.
Because it is my priority to make this package better by relaxing requirements and using conda instead of pip.

### mixture of pip and conda

I think I see what you mean, when you install the package with pip in the conda environment Sandpyper automatically installs (again) the packages and uses pip.
Is this the issue you are referring to?

Because if the conda environment `sandpyper_env `is created by copying and pasting the requirements in the `README`, when you install `sandpyper 1.2.0` (latest) using pip, you should have already all the right packages installed and this should only be a kind of check.
If for some reason your environment has some required packages that differ from the versions specified in both the `conda create environment ... `(`README`) and in the `requirements.txt,` then it uses pip to install those. And that is causing issues.

In my case though this doesn't happen. See the image below.
After I created the `sandpyper_env` (in this case renamed sandpyper_env2), I install sandpyper using pip and this is what I mean by "check".

![image](https://user-images.githubusercontent.com/44853536/135364887-0ae8462b-bb79-49a3-a0c4-fdffe59f3f29.png)

But if that is not what happens to you, there is something weird going on. __I am looking into it right now.__

### To run the notebooks., the instructions are inadequate for a novice user

You are right, I hadn't thought about this!
Actually this is a really good point and I would love to hear your suggestion on how to solve this. Even once the package will be in conda, to run the notebooks users will need to download the repo anyway to get the test_data.rar (which by the way I forgot to include it in its compressed format in the repo).
__I have 2 ideas:__
1. Host the test_data.rar somewhere else easy to download
2. Use binder to run the notebooks online

option 1 is easy and I can do it quickly. Option 2 is much better but I must first learn how to use binder and then do it, which might take a bit. But looking at [this](https://mybinder.org/) seems easy.
Anyway, I now update the `README` and the notebooks to explaining how to get the test_data.rar.

### In the first notebook, "1 - Introduction and data preparation", you don't explain where 'demo_data.rar' is located.

Sorry, definitely something I will explain at the beginning of the notebook.
As per the propeller platform, the data is freely __accessible but not downloadable at the moment__. You can access it, visualize it, make measurements and see changes in all 20 sites in the past 3 years but as of today, it is not downloadable. I think it will eventually be but I have to check with my supervisor.
I see the confusion here, in the paper there is written __freely accessible__ which can be misleading.

> More than 220 3D datasets are already __freely accessible to anyone__ via a user friendly web-platform to share and communicate information, promote coastal awareness, build knowledge and further increase the impact of our efforts.

I update the README and the notebook and find a way to make the `test_data.rar `at least downloadable.
Thanks again!
Nick
