### 2021-10-03 d7f0632
test no install requires for PyPI
