+++ setup.py
@@ -32,7 +32,7 @@ setup(
     python_requires='>=3.5',
     classifiers=[
         'Development Status :: 2 - Pre-Alpha',
-        'Intended Audience :: Developers',
+        'Intended Audience :: Science/Research',
         'License :: OSI Approved :: MIT License',
         'Natural Language :: English',
         'Programming Language :: Python :: 3',
@@ -42,13 +42,13 @@ setup(
         'Programming Language :: Python :: 3.8',
     ],
     description="Tools for automatic UAV-SfM beach volumetric and behavioural analysis.",
-    install_requires=install_requires,
+    #install_requires=install_requires,
     dependency_links=dependency_links,
     license="MIT license",
     long_description=readme,
     long_description_content_type='text/markdown',
     include_package_data=True,
-    keywords='sandpyper',
+    keywords='sandpyper beach analysis SfM raster volume profiles UAV drone monitoring',
     name='sandpyper',
     packages=find_packages(include=['sandpyper', 'sandpyper.*']),
     setup_requires=setup_requirements,
