+++ setup.py
@@ -55,6 +55,6 @@ setup(
     test_suite='tests',
     tests_require=test_requirements,
     url='https://github.com/npucino/sandpyper',
-    version='1.2.0',
+    version='1.2.1',
     zip_safe=False,
 )
