+++ sandpyper/__init__.py
@@ -1,6 +1,6 @@
 
 """Profile module for sandpyper."""
 
-__version__ = "1.2.0"
+__version__ = "1.2.1"
 __author__ = """Nicolas Pucino"""
 __email__ = 'npucino@deakin.edu.au'
