+++ setup.py
@@ -29,16 +29,12 @@ test_requirements = [ ]
 setup(
     author="Nicolas Pucino",
     author_email='npucino@deakin.edu.au',
-    python_requires='>=3.5',
+    python_requires='>=3.8',
     classifiers=[
         'Development Status :: 2 - Pre-Alpha',
         'Intended Audience :: Science/Research',
         'License :: OSI Approved :: MIT License',
         'Natural Language :: English',
-        'Programming Language :: Python :: 3',
-        'Programming Language :: Python :: 3.5',
-        'Programming Language :: Python :: 3.6',
-        'Programming Language :: Python :: 3.7',
         'Programming Language :: Python :: 3.8',
     ],
     description="Tools for automatic UAV-SfM beach volumetric and behavioural analysis.",
