### 2021-10-04 c56c785
test minimum versions except sklearn
