+++ tests/test_sandpyper.py
@@ -29,7 +29,7 @@ loc_specs={'mar':{'thresh':6,
        'min_points':6}}
 
 loc_search_dict = {'leo': ['St', 'Leonards', 'leonards', 'leo'], 'mar': ['Marengo', 'marengo', 'mar'] }
-crs_dict_string = {'mar': {'init': 'epsg:32754'}, 'leo':{'init': 'epsg:32755'} }
+crs_dict_epsg = {'mar': 32754}, 'leo':32755 }
 
 labels=["Undefined", "Small", "Medium", "High", "Extreme"]
 
@@ -146,7 +146,7 @@ class TestCreateProfiles(unittest.TestCase):
                            sampling_step=20,
                            tick_length=50,
                            location='leo',
-                           crs=crs_dict_string['leo'],
+                           crs=crs_dict_epsg['leo'],
                            side='both'
                           )
         self.assertEqual(transects_leo.shape, (59, 3))
@@ -160,7 +160,7 @@ class TestCreateProfiles(unittest.TestCase):
                        sampling_step=20,
                        tick_length=50,
                        location='mar',
-                       crs=crs_dict_string['mar'],
+                       crs=crs_dict_epsg['mar'],
                        side='both'
                       )
 
@@ -174,7 +174,7 @@ class TestCreateProfiles(unittest.TestCase):
             create_transects(self.leo_shoreline,
                            sampling_step=0,
                            tick_length=100,
-                           location='leo' ,crs=crs_dict_string['leo'],
+                           location='leo' ,crs=crs_dict_epsg['leo'],
                            side='both'
                           )
 
@@ -182,13 +182,13 @@ class TestCreateProfiles(unittest.TestCase):
         right=create_transects(self.mar_shoreline,
                            sampling_step=150,
                            tick_length=50,
-                           location='mar' ,crs=crs_dict_string['mar'],
+                           location='mar' ,crs=crs_dict_epsg['mar'],
                            side='right'
                           )
         left=create_transects(self.mar_shoreline,
                                    sampling_step=150,
                                    tick_length=50,
-                                   location='mar' ,crs=crs_dict_string['mar'],
+                                   location='mar' ,crs=crs_dict_epsg['mar'],
                                    side='left'
                                   )
         self.assertTrue(right.touches(left).all())
@@ -216,7 +216,7 @@ class TestProfileSet(unittest.TestCase):
                         transects_spacing=transects_spacing,
                         loc_codes=loc_codes,
                         loc_search_dict=loc_search_dict,
-                        crs_dict_string=crs_dict_string,
+                        crs_dict_epsg=crs_dict_epsg,
                         check="all")
 
 
@@ -309,9 +309,9 @@ class TestProfileSet(unittest.TestCase):
         test_slice_last=test_slice.query(f"tr_id=={max(test_slice.tr_id.unique())}").copy()
         test_slice_rand=test_slice.query(f"tr_id=={np.random.randint(min(test_slice.tr_id.unique()),max(test_slice.tr_id.unique()))}").iloc[1:,:].copy()
 
-        pts_spacing_0=np.nanmean(test_slice_tr0.coordinates.set_crs(32754).distance(test_slice_tr0.next_point.set_crs(32754)))
-        pts_spacing_last=np.nanmean(test_slice_last.coordinates.set_crs(32754).distance(test_slice_last.next_point.set_crs(32754)))
-        pts_spacing_rand=np.nanmean(test_slice_rand.coordinates.set_crs(32754).distance(test_slice_rand.next_point.set_crs(32754)))
[…]
