+++ sandpyper/common.py
@@ -15,6 +15,7 @@ from shutil import move, copy
 import numpy as np
 import pandas as pd
 import geopandas as gpd
+from pyproj import CRS
 from geopandas.tools import overlay
 import seaborn as sb
 from tqdm.notebook import tqdm_notebook as tqdm
@@ -59,11 +60,17 @@ import richdem as rd
 from statsmodels.api import qqplot
 
 from pysal.lib import weights
-import pysal.explore.esda.moran as moran
-from pysal.explore.esda.util import fdr
+#import pysal.explore.esda.moran as moran
+#from pysal.explore.esda.util import fdr
+from esda import moran
+from esda.util import fdr
 
-from pysal.explore.giddy.markov import Markov
-from pysal.lib.weights import DistanceBand, Queen, higher_order
+#from pysal.explore.giddy.markov import Markov
+from giddy.markov import Markov
+
+#from pysal.lib.weights import DistanceBand, Queen, higher_order
+from libpysal.weights import  DistanceBand, Queen, higher_order
+
 from pysal.viz.mapclassify import (EqualInterval,
                                    FisherJenks,
                                    HeadTailBreaks,
@@ -1477,7 +1484,7 @@ def plot_lod_normality_check(multitemp_data, lod_df, details_table, locations,al
             ax1.grid(axis='y')
             ax1.grid(b=None,axis='x')
 
-def get_rbcd_transect(df_labelled, loc_specs, reliable_action, dirNameTrans, labels_order, loc_codes, crs_dict_string):
+def get_rbcd_transect(df_labelled, loc_specs, reliable_action, dirNameTrans, labels_order, loc_codes, crs_dict_epsg):
     """It computes the r-BCDs at the transect level, based on the timeseries of elevation change magnituteds across the beachface dataset (markov_tag dataframe).
 
     Args:
@@ -1487,7 +1494,7 @@ def get_rbcd_transect(df_labelled, loc_specs, reliable_action, dirNameTrans, lab
         dirNameTrans (str): Path of the directory containing the transects (geopackages, .gpkg).
         labels_order (list): Order of labels (magnitude of change) to be preserved.
         loc_codes (list): List of strings containing location codes.
-        crs_dict_string (dict): Dictionary storing location codes as key and crs information as values, in dictionary form.
+        crs_dict_epsg (dict): Dictionary storing location codes as key and crs information as EPSG code (int).
 
     Returns:
        rbcd_transects (gpd.GeoDataFrame): A GeoDataFrames containing the steady-state distribution of each transect.
@@ -1625,7 +1632,7 @@ def get_rbcd_transect(df_labelled, loc_specs, reliable_action, dirNameTrans, lab
                 ).geometry
 
         ss_transects_idx_loc = gpd.GeoDataFrame(
-                        merged_erodepo, geometry="geometry", crs=crs_dict_string[loc]
+                        merged_erodepo, geometry="geometry", crs=CRS.from_epsg(crs_dict_epsg[loc])
                     )
         ss_transects_idx=pd.concat([ss_transects_idx_loc,ss_transects_idx], ignore_index=True)
 
@@ -1816,7 +1823,7 @@ def sensitivity_tr_rbcd(ProfileDynamics,
                       dirNameTrans=ProfileDynamics.ProfileSet.dirNameTrans,
                       labels_order=ProfileDynamics.tags_order,
                       loc_codes=ProfileDynamics.ProfileSet.loc_codes,
-                      crs_dict_string=ProfileDynamics.ProfileSet.crs_dict_string)
+                      crs_dict_epsg=ProfileDynamics.ProfileSet.crs_dict_epsg)
 
                 ss_transects_idx['thresh']=i[1]
                 ss_transects_idx['min_pts']=i[0]
[…]
