### 2021-10-04 98f87f0
less strict sklearn
