+++ requirements.txt
@@ -1,8 +1,7 @@
-geopandas
+geopandas=0.8
 pysal
 rasterio
 richdem
-scikit-image
 scikit-learn=0.24
 seaborn
 tqdm
