### 2021-10-04 991dc74
back to strict, with code mods
