### 2021-10-04 9bbecbe
strict reqs with new geopandas code and old pysal importts
