+++ sandpyper/common.py
@@ -1484,7 +1484,7 @@ def plot_lod_normality_check(multitemp_data, lod_df, details_table, locations,al
             ax1.grid(axis='y')
             ax1.grid(b=None,axis='x')
 
-def get_rbcd_transect(df_labelled, loc_specs, reliable_action, dirNameTrans, labels_order, loc_codes, crs_dict_epsg):
+def get_rbcd_transect(df_labelled, loc_specs, reliable_action, dirNameTrans, labels_order, loc_codes, crs_dict_string):
     """It computes the r-BCDs at the transect level, based on the timeseries of elevation change magnituteds across the beachface dataset (markov_tag dataframe).
 
     Args:
@@ -1494,7 +1494,7 @@ def get_rbcd_transect(df_labelled, loc_specs, reliable_action, dirNameTrans, lab
         dirNameTrans (str): Path of the directory containing the transects (geopackages, .gpkg).
         labels_order (list): Order of labels (magnitude of change) to be preserved.
         loc_codes (list): List of strings containing location codes.
-        crs_dict_epsg (dict): Dictionary storing location codes as key and crs information as values, in dictionary form.
+        crs_dict_string (dict): Dictionary storing location codes as key and crs information as values, in dictionary form.
 
     Returns:
        rbcd_transects (gpd.GeoDataFrame): A GeoDataFrames containing the steady-state distribution of each transect.
@@ -1632,7 +1632,7 @@ def get_rbcd_transect(df_labelled, loc_specs, reliable_action, dirNameTrans, lab
                 ).geometry
 
         ss_transects_idx_loc = gpd.GeoDataFrame(
-                        merged_erodepo, geometry="geometry", crs=CRS.from_epsg(crs_dict_epsg[loc])
+                        merged_erodepo, geometry="geometry", crs=CRS.from_epsg(crs_dict_string[loc])
                     )
         ss_transects_idx=pd.concat([ss_transects_idx_loc,ss_transects_idx], ignore_index=True)
 
@@ -1823,7 +1823,7 @@ def sensitivity_tr_rbcd(ProfileDynamics,
                       dirNameTrans=ProfileDynamics.ProfileSet.dirNameTrans,
                       labels_order=ProfileDynamics.tags_order,
                       loc_codes=ProfileDynamics.ProfileSet.loc_codes,
-                      crs_dict_epsg=ProfileDynamics.ProfileSet.crs_dict_epsg)
+                      crs_dict_string=ProfileDynamics.ProfileSet.crs_dict_string)
 
                 ss_transects_idx['thresh']=i[1]
                 ss_transects_idx['min_pts']=i[0]
@@ -2219,7 +2219,7 @@ def classify_labelk(labelled_dataset,l_dicts, cluster_field='label_k', fill_clas
 def cleanit(to_clean, l_dicts, cluster_field='label_k', fill_class='sand',
             watermasks_path=None, water_label='water',
             shoremasks_path=None, label_corrections_path=None,
-            crs_dict_epsg=None,
+            crs_dict_string=None,
            geometry_field='coordinates'):
 
     print("Reclassifying dataset with the provided dictionaries." )
@@ -2239,8 +2239,8 @@ def cleanit(to_clean, l_dicts, cluster_field='label_k', fill_class='sand',
             print(f"Label corrections provided in CRS: {label_corrections.crs}")
 
             # validate label correction polygons
-            #if crs_dict_epsg != None:
-            #    check_overlaps_poly_label(label_corrections,to_clean_classified,crs_dict_epsg)
+            #if crs_dict_string != None:
+            #    check_overlaps_poly_label(label_corrections,to_clean_classified,crs_dict_string)
             #else:
[…]
