### 2021-10-04 #3666 @crvernon
@dbuscombe-usgs and @npucino - I noticed that all tasks have been checked.  Are you still collaborating on fixes for this review or have all of your inquiries been addressed @dbuscombe-usgs ?

Also, I noticed that the following related issues are still open:

- [ ] https://github.com/npucino/sandpyper/issues/6
- [ ] https://github.com/npucino/sandpyper/issues/8
