### 2021-10-04 #3666 @dbuscombe-usgs
Hi @cvernon, I am satisfied - I approve this submission

I will still work with @npucino to help test remaining issues with conda packaging
