+++ sandpyper/common.py
@@ -1632,7 +1632,7 @@ def get_rbcd_transect(df_labelled, loc_specs, reliable_action, dirNameTrans, lab
                 ).geometry
 
         ss_transects_idx_loc = gpd.GeoDataFrame(
-                        merged_erodepo, geometry="geometry", crs=CRS.from_epsg(crs_dict_string[loc])
+                        merged_erodepo, geometry="geometry", crs=crs_dict_string[loc]
                     )
         ss_transects_idx=pd.concat([ss_transects_idx_loc,ss_transects_idx], ignore_index=True)
 
@@ -2271,7 +2271,7 @@ def cleanit(to_clean, l_dicts, cluster_field='label_k', fill_class='sand',
                         selection=data_in[data_in.coordinates.intersects(row.geometry)]
 
                         if selection.shape[0]==0:
-                            selection=data_in[data_in.to_crs(CRS.from_epsg(crs_dict_string[loc])).coordinates.intersects(row.geometry)]
+                            selection=data_in[data_in.to_crs(crs_dict_string[loc]).coordinates.intersects(row.geometry)]
                         else:
                             pass
                         selection["finetuned_label"]=new_class
@@ -2345,7 +2345,7 @@ def cleanit(to_clean, l_dicts, cluster_field='label_k', fill_class='sand',
                     selection=subset_data[subset_data.geometry.intersects(subset_masks.geometry)]
                     if selection.shape[0]==0:
                         print(f"Reprojecting")
-                        selection=subset_data[subset_data.geometry.intersects(subset_masks.to_crs(CRS.from_epsg(crs_dict_string[loc])).geometry.any())]
+                        selection=subset_data[subset_data.geometry.intersects(subset_masks.to_crs(crs_dict_string[loc]).geometry.any())]
                     else:
                         pass
 
@@ -2414,7 +2414,7 @@ def cleanit(to_clean, l_dicts, cluster_field='label_k', fill_class='sand',
                 if in_shore.shape[0]>=1:
                     pass
                 else:
-                    in_shore=loc_selection[loc_selection.geometry.intersects(shore.to_crs(CRS.from_epsg(crs_dict_string[loc])).geometry.any())]
+                    in_shore=loc_selection[loc_selection.geometry.intersects(shore.to_crs(crs_dict_string[loc]).geometry.any())]
 
                 print(f"Removing {loc_selection.shape[0] - in_shore.shape[0]} pts falling outside provided shore polygones.")
                 inshore_cleaned=pd.concat([in_shore,inshore_cleaned], ignore_index=True)
@@ -4197,7 +4197,7 @@ def error_from_gt(
     Returns:
         pd.DataFrame: Dataframe containing the distances from groundtruth at each timestep.
     """
-    crs = CRS.from_epsg(crs_dict_string[loc])
+    crs = crs_dict_string[loc]
 
     if os.path.isfile(baseline_mode):
         print("Fixed baseline mode selected.")
@@ -4529,7 +4529,7 @@ def tidal_correction(
         gt_shores = cs_shores.query(
             f"location=='{location}' & raw_date in @list_dates"
         )  # get groudtruths shores
-        crs = CRS.from_epsg(crs_dict_string[loc])  # get crs of location
+        crs = crs_dict_string[loc]  # get crs of location
 
         if shores_to_corr.crs != crs:
             shores_to_corr = shores_to_corr.to_crs(crs)
@@ -5175,7 +5175,7 @@ def save_slope_corr_files(
         print(f"Working on {location} .")
[…]
