### 2021-10-05 9f7efa2
test semi-relaxed requirements
