+++ requirements.txt
@@ -5,7 +5,7 @@ tqdm
 pysal==2.1
 rasterio
 richdem
-scikitimage=0.18
-scikitlearn=0.24
+scikit-image=0.18
+scikit-learn=0.24
 seaborn
 fuzzywuzzy
