### 2021-10-05 50686c3
added scipy 1.6.0
