+++ requirements.txt
@@ -7,3 +7,4 @@ scikit-learn==0.24.1
 seaborn
 tqdm
 fuzzywuzzy
+scipy==1.6.0
