### 2021-10-05 #3666 @npucino
Hi guys, **thanks again** for you extremely valuable comments and contributions!

Yesterday I was trying the .yml option with different degrees of version requirements and still getting some tests to fail.
Right now I am also testing different versions of requirements to see if I can relax at least part of the requirements and isolate the issue.

But,
using the strict version requirements, tests do not fail so for now, I will stick with this sub-optimal solution.

By the way @dbuscombe-usgs, I added` test_data.rar` in the repo and updated `README` and notebooks with direct link for downloading the data. I am now trying the last few versions of yml and the I update the `README` so that users will use the `conda env create -f sandpyper_env.yml` option.

I never learnt so much in a review process. Thanks a lot @dbuscombe-usgs @chrisleaman @crvernon .
Nick
