### 2021-10-05 #3666 @npucino
### Sorry, last quick update!

Actually, I am close to nail down this weird random generator issue.
Apparently something happened between numpy 1.20 and 1.21 with their [default  random number generator](https://numpy.org/devdocs/reference/random/upgrading-pcg64.html#upgrading-pcg64).
I really start to think that is numpy that causes my tests to fail behind the curtains, affecting sklearn KMeans function, because I noticed that if I use numpy 1.20 it works, if I use 1.21 tests do not pass.
Therefore, the only requirement I might  need is numpy 1.20 but I still have to try it.
And, numpy is so fundamental that it is not a long-term solution at all, but, I will get there!

Cheers!
