+++ sandpyper/common.py
@@ -60,16 +60,17 @@ import richdem as rd
 from statsmodels.api import qqplot
 
 from pysal.lib import weights
-import pysal.explore.esda.moran as moran
-from pysal.explore.esda.util import fdr
-#from esda import moran
-#from esda.util import fdr
 
-from pysal.explore.giddy.markov import Markov
-#from giddy.markov import Markov
+#import pysal.explore.esda.moran as moran
+#from pysal.explore.esda.util import fdr
+from esda import moran
+from esda.util import fdr
 
-from pysal.lib.weights import DistanceBand, Queen, higher_order
-#from libpysal.weights import  DistanceBand, Queen, higher_order
+#from pysal.explore.giddy.markov import Markov
+from giddy.markov import Markov
+
+#from pysal.lib.weights import DistanceBand, Queen, higher_order
+from libpysal.weights import  DistanceBand, Queen, higher_order
 
 from pysal.viz.mapclassify import (EqualInterval,
                                    FisherJenks,
@@ -1632,7 +1633,7 @@ def get_rbcd_transect(df_labelled, loc_specs, reliable_action, dirNameTrans, lab
                 ).geometry
 
         ss_transects_idx_loc = gpd.GeoDataFrame(
-                        merged_erodepo, geometry="geometry", crs=crs_dict_string[loc]
+                        merged_erodepo, geometry="geometry", crs=CRS.from_epsg(CRS.from_epsg(crs_dict_string[loc]))
                     )
         ss_transects_idx=pd.concat([ss_transects_idx_loc,ss_transects_idx], ignore_index=True)
 
@@ -2271,7 +2272,7 @@ def cleanit(to_clean, l_dicts, cluster_field='label_k', fill_class='sand',
                         selection=data_in[data_in.coordinates.intersects(row.geometry)]
 
                         if selection.shape[0]==0:
-                            selection=data_in[data_in.to_crs(crs_dict_string[loc]).coordinates.intersects(row.geometry)]
+                            selection=data_in[data_in.to_crs(CRS.from_epsg(crs_dict_string[loc])).coordinates.intersects(row.geometry)]
                         else:
                             pass
                         selection["finetuned_label"]=new_class
@@ -2345,7 +2346,7 @@ def cleanit(to_clean, l_dicts, cluster_field='label_k', fill_class='sand',
                     selection=subset_data[subset_data.geometry.intersects(subset_masks.geometry)]
                     if selection.shape[0]==0:
                         print(f"Reprojecting")
-                        selection=subset_data[subset_data.geometry.intersects(subset_masks.to_crs(crs_dict_string[loc]).geometry.any())]
+                        selection=subset_data[subset_data.geometry.intersects(subset_masks.to_crs(CRS.from_epsg(crs_dict_string[loc])).geometry.any())]
                     else:
                         pass
 
@@ -2414,7 +2415,7 @@ def cleanit(to_clean, l_dicts, cluster_field='label_k', fill_class='sand',
                 if in_shore.shape[0]>=1:
                     pass
                 else:
-                    in_shore=loc_selection[loc_selection.geometry.intersects(shore.to_crs(crs_dict_string[loc]).geometry.any())]
+                    in_shore=loc_selection[loc_selection.geometry.intersects(shore.to_crs(CRS.from_epsg(crs_dict_string[loc])).geometry.any())]
 
                 print(f"Removing {loc_selection.shape[0] - in_shore.shape[0]} pts falling outside provided shore polygones.")
[…]
