+++ requirements.txt
@@ -1,6 +1,7 @@
 geopandas
 numpy==1.20.1
 rasterio
+pysal
 richdem
 scikit-image
 scikit-learn
