+++ requirements.txt
@@ -4,7 +4,7 @@ rasterio
 pysal
 richdem
 scikit-image
-scikit-learn
+scikit-learn==0.24.1
 seaborn
 tqdm
 fuzzywuzzy
