+++ requirements.txt
@@ -1,4 +1,5 @@
 geopandas
+pandas==1.2.2
 numpy==1.20.1
 rasterio
 pysal
