### 2021-10-05 9eb4b4b
back to strict reqs
