### 2021-10-05 c10dce1
Revert "restored strict reqs"
This reverts commit 006c6d5987632b3a0d07f1df7cff77b536bf844e.
