+++ requirements.txt
@@ -1,14 +1,10 @@
 geopandas==0.8.2
-matplotlib==3.3.4
-numpy==1.20.1
-pandas==1.2.2
 pysal==2.1
 rasterio==1.2.0
 richdem==0.3.4
 scikit-image==0.18.1
 scikit-learn==0.24.1
-scipy==1.6.0
-seaborn==0.11.1
-tqdm==4.56.2
-pooch==1.4.0
+seaborn
+tqdm
 fuzzywuzzy
+scipy==1.6.0
