### 2021-10-06 9ab1fda
restored crs_dict init form
