+++ .github/workflows/build.yml
@@ -15,8 +15,8 @@ jobs:
             fail-fast: false
             matrix:
                 config:
-                    - { os: windows-latest, py: "3.8" }
-                    - { os: macOS-latest, py: "3.8" }
+                    - { os: windows-latest, py: "3.9" }
+                    - { os: macOS-latest, py: "3.9" }
                     #- { os: ubuntu-latest, py: "3.8" }
                     #- { os: ubuntu-latest, py: "3.9" }
         env:
