### 2021-10-09 c1ca746
-try pre Kmeans classified data
