+++ tests/test_sandpyper.py
@@ -347,10 +347,10 @@ class TestProfileSet(unittest.TestCase):
     def test_008_correction(self):
 
         #self.assertIsInstance(self.P.profiles.label_k[0],  np.int32)
-        self.assertEqual(self.check_watermask_mar_pre,(219, 14))
-        self.assertEqual(self.check_watermask_leo_pre,(300, 14))
-        self.assertEqual(self.check_watermask_leo_post,(24, 15))
-        self.assertEqual(self.check_watermask_mar_post,(219, 15))
+        self.assertEqual(self.check_watermask_mar_pre,(219, 15))
+        self.assertEqual(self.check_watermask_leo_pre,(300, 15))
+        self.assertEqual(self.check_watermask_leo_post,(24, 16))
+        self.assertEqual(self.check_watermask_mar_post,(219, 16))
 
         self.assertTrue(self.check_no_sand_mar_postclean.pt_class.unique()=='no_sand')
         self.assertTrue(self.check_no_sand_leo_postclean.pt_class.unique()=='no_sand')
