### 2021-10-09 606aebd
mods to super relaxed reqs with pre KMeans data
