### 2021-10-09 94e5ae3
revert to strict reqs
