+++ .github/workflows/build.yml
@@ -16,9 +16,11 @@ jobs:
             matrix:
                 config:
                     - { os: windows-latest, py: "3.9" }
+                    - { os: windows-latest, py: "3.8" }
                     - { os: macOS-latest, py: "3.9" }
-                    #- { os: ubuntu-latest, py: "3.8" }
-                    #- { os: ubuntu-latest, py: "3.9" }
+                    - { os: macOS-latest, py: "3.8" }
+                    - { os: ubuntu-latest, py: "3.8" }
+                    - { os: ubuntu-latest, py: "3.9" }
         env:
             SDKROOT: /Library/Developer/CommandLineTools/SDKs/MacOSX.sdk
 
