### 2021-10-09 a3d54fb
Bump version: 1.2.1 → 1.2.2
