+++ setup.py
@@ -51,6 +51,6 @@ setup(
     test_suite='tests',
     tests_require=test_requirements,
     url='https://github.com/npucino/sandpyper',
-    version='1.2.1',
+    version='1.2.2',
     zip_safe=False,
 )
