### 2021-10-09 c10cd29
eliminated useless temp data
