### 2021-10-11 8a0687c
JOSS paper https://github.com/openjournals/joss-reviews/issues/3666#issuecomment-940053558
