+++ paper/paper.md
@@ -32,28 +32,28 @@ bibliography: paper.bib
 # Introduction
 
 Coastal zones host 40% of the world population [@Martinez2007-ux]  with continued expected growth to be focused in least developed countries [@Neumann2015-ph]. Sandy beaches, amongst other ecoservices [@Barbier2011-qp], protect inland assets from coastal erosion, dissipating stormy waves energy on their shores. Mitigating beach erosion typically involves the establishment of topographic monitoring programs in key locations (erosional hotspots) to quantify beach dynamics, erosion/deposition volumes, recovery times and model coastal resilience or risk to erosion [@Turner2016-dv, @Ruessink2019-yl]. High temporal and spatial resolution topographic data is ideal, but expensive with most of the ordinary beach surveying methods.
-Unoccupied Aerial Vehicles (UAVs) and Structure from Motion algorithms (UAV-SfM) are emerging as ideal platform and methodology to obtain cost-effective high-quality beach topographic data (as Digital Surface Models, DSMs) [@Goncalves2015-qr] at the mesoscale, which is an appropriate spatiotemporal scale for coastal managers to work with [@Thom2018-kf]. Consequently, researchers already use UAV-SfM to monitor beach dynamics around the world [@Brunier2016-hq, @Jaud2019-rp], but it has been limited so far to a few sites and a few multitemporal dates. However more recently, UAV-SfM technology is being increasingly used for wider-scale and longer-term monitoring projects.
+Unoccupied Aerial Vehicles (UAVs) and Structure from Motion algorithms (UAV-SfM) are emerging as ideal platform and methodology to obtain cost-effective high-quality beach topographic data [as Digital Surface Models, DSMs; @Goncalves2015-qr] at the mesoscale, which is an appropriate spatiotemporal scale for coastal managers to work with [@Thom2018-kf]. Consequently, researchers already use UAV-SfM to monitor beach dynamics around the world [@Brunier2016-hq, @Jaud2019-rp], but it has been limited so far to a few sites and a few multitemporal dates. However more recently, UAV-SfM technology is being increasingly used for wider-scale and longer-term monitoring projects.
 For instance in Victoria (Australia), a citizen-science UAV-SfM monitoring program mobilises more than 150 volunteers to fly UAVs on 15 sites every six weeks during the last three years. To date, volunteers flew 350 times, enabling the creation of a DSM and an orthophoto per survey (uncompressed file sizes from 5-100 Gb each). Citizen scientists and SfM are generating an unprecedented archive of imagery which can be reliably used to monitor high-frequency sandy beach volumetric dynamics and behaviors [@Pucino2021-ox].
 
 # Statement of Need
 
-A drawback of using UAV-SfM for beach monitoring is that due to UAV regulations, flight altitude is often limited to around 80-120 m above ground, which means that the ground sampling distance of consumer-grade UAVs is sub-decimeter, resulting in very high resolution and large imagery files, especially for beach surveys exceeding the 20 ha coverage. Although managing tens of large rasters with geographic information systems such as QGIS [@QGIS_software] or ESRI ArcGIS is technically feasible, handling tens to hundreds of such files within large monitoring projects quickly becomes impractical.
-Moreover, in coastal management, erosion assessments from multitemporal DSMs are usually approached by raster subtraction (also known as DEM of difference method, see [@Lane2000-gs]), which is a process to compute elevation difference between time intervals by subtracting the elevation value of each cell in the pre and post rasters. Raster-based operations with full-resolution UAV-SfM imagery becomes very time consuming with important computing power and memory needs that can limit their feasibility.
[…]
