### 2021-10-11 087c661
images references added https://github.com/openjournals/joss-reviews/issues/3666#issuecomment-940053558
