+++ setup.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 1.2.2
+current_version = 1.3.0
 commit = True
 tag = True
 
@@ -19,3 +19,4 @@ exclude = docs
 max-line-length = 88
 
 [aliases]
+
