+++ ML and SA/ML and SA.md
@@ -1,0 +1,157 @@
+# Methods for sand labelling and classification
+
+![im](images/transects_classes.png)
+
+
+As the Structure from Motion pipeline assumes tie-points to be fixed during multi-view stereo-pairing, the swash zone created obvious artifacts that would interfere with an accurate representation of the beach topography. Its extent and location is highly dynamic, due to the combined effect of tidal and water level variations, planimetric beach adjustments and variable survey times.  
+Also, dune vegetation, beach wracks and anything that was neither sand nor a fixed feature (like rock walls or fences) on the beach needed to be removed to exclusively obtain sand-specific volumetric changes and accurate profiles evolution.
+For this reason, Sandpyper uses a machine learning algorithm to semi-automatically classify sand and non-sand points (objects, seaweed wracks, vegetation), facilitating the cleaning of big datasets.
+
+In order to classify sand and no-sand points sampled across transects, Sandpyper facilitates the use of Silhouette Analysis (SA) method to propose a sub-optimal __number of clusters k__ to be used by the unsupervised KMeans clustering algorithm. Following the unsupervised procedure, Sandpyper helps in further cleaning the dataset by providing it with correction polygons (in case the user wants to fine-tune the KMeans assigned labels), watermasks (to further mask out water and swash) and shoremasks (to clip the area of study to not too far landward from the foredune, if existent).
+
+A typical Sandpyper pipeline consists of:
+
+1. Automated iterative Silhouette Analysis with inflexion point search at the site level, implemented with `get_sil_location()` function
+2. Inflexion points search, implemented with `get_opt_k()` function
+3. Run KMeans with the identified sub-optimal k, implemented with the `ProfileSet.kmeans_sa()` method.
+4. Visual assessment of KMeans generated labels and creation of class dictionaries and correction polygons in a GIS
+5. Semi-automated classification and cleaning, implemented with `ProfileSet.cleanit()` method.
+
+
+## Unsupervised clustering with KMeans
+KMeans is one of the most used, versatile and readily available clustering algorithms in data mining ([Wu et al., 2007](http://dx.doi.org/10.1007/s10115-007-0114-2)) and
+showed its value on partitioning SfM-MVS derived coastal point clouds ([Callow et al., 2018](http://dx.doi.org/10.1007/s10115-007-0114-2)).
+For a general overview of clustering algorithms with specific attention on KMeans algorithm,
+please refer to [Jain (2010)](https://www.sciencedirect.com/science/article/abs/pii/S0167865509002323?via%3Dihub).
+
+Sandpyper implements Elkan's KMeans algorithm ([Elkan, 2003](https://www.aaai.org/Papers/ICML/2003/ICML03-022.pdf)) with triangle inequality acceleration via [sklearn](https://scikit-learn.org/stable/modules/generated/sklearn.cluster.KMeans.html) package, with the following parameters:
+1. initial cluster centers selected with k-means++
+2. 300 maximum iterations per run
+3. inertial tolerance of 0.0001
+4. precomputed distances if n_samples * n_clusters < 12 million
+
+Have a look at this short and easy video about the intuition of KMeans.<br>
+![im](images/kmeans.JPG)
+[Watch here](https://www.youtube.com/watch?v=4b5d3muPQmA)
+
+
+## Silhouette Analysis (SA)
+
+The Silhouette Analysis (SA) method is a simple graphical and analytical method to measure how tight and compact the clusters are (overall), while also indicating how well each observation fits (silhouette coefficient) within the assigned partition or cluster ([Rousseeuw, 1987](https://www.sciencedirect.com/science/article/pii/0377042787901257?via%3Dihub)).
+
[…]
