+++ sandpyper/sandpyper.md
@@ -1,0 +1,1 @@
+::: sandpyper.sandpyper
