+++ installation/installation.md
@@ -1,0 +1,23 @@
+# Installation
+
+## Stable release
+
+To install sandpyper, run this command in your terminal:
+
+```
+pip install sandpyper
+```
+
+This is the preferred method to install sandpyper, as it will always install the most recent stable release.
+
+If you don't have [pip](https://pip.pypa.io) installed, this [Python installation guide](http://docs.python-guide.org/en/latest/starting/installation/) can guide you through the process.
+
+## From sources
+
+The sources for sandpyper can be downloaded from the Github repo.
+
+You can clone the public repository:
+
+```
+git clone git://github.com/npucino/sandpyper
+```

[…]
